+++ app-shell/.dockerignore
