+++ app-shell/Dockerfile
