+++ app-shell/package.json
@@ -1,18 +1,18 @@
 {
-  "name": "my-app",
+  "name": "app-shell",
   "version": "1.0.0",
   "description": "",
   "keywords": [
     "piral"
   ],
   "dependencies": {
-    "piral": "^1.3.2",
-    "piral-base": "^1.3.2",
-    "piral-core": "^1.3.2",
+    "piral": "1.5.2",
+    "piral-base": "1.5.2",
+    "piral-core": "1.5.2",
     "react": "^18",
     "react-dom": "^18",
-    "react-router": "^6.18.0",
-    "react-router-dom": "^6.18.0"
+    "react-router": "^5",
+    "react-router-dom": "^5"
   },
   "scripts": {
     "start": "piral debug",
@@ -30,13 +30,13 @@
     ]
   },
   "devDependencies": {
-    "@types/node": "^20.8.10",
+    "@types/node": "latest",
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "@types/react-router": "^5",
     "@types/react-router-dom": "^5",
-    "piral-cli": "^1.3.2",
-    "piral-cli-webpack5": "^1.3.2",
+    "piral-cli": "1.5.2",
+    "piral-cli-webpack5": "1.5.2",
     "typescript": "latest"
   }
 }
