+++ app-shell/piral.json
@@ -1,5 +1,6 @@
 {
   "$schema": "https://docs.piral.io/schemas/piral-v0.json",
+  "isolation": "modern",
   "pilets": {
     "files": [],
     "scripts": {},
