--- my-app/pilets/product-pilet/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "product-pilet",
-  "version": "1.0.0",
-  "description": "",
-  "keywords": [
-    "pilet"
-  ],
-  "devDependencies": {
-    "@types/node": "latest",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "my-app": "file:..\\dist\\emulator\\my-app-1.0.0.tgz",
-    "piral-cli": "1.2.0",
-    "piral-cli-webpack5": "1.2.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-router": "5.3.4",
-    "react-router-dom": "5.3.4",
-    "tslib": "2.6.2",
-    "typescript": "latest"
-  },
-  "scripts": {
-    "start": "pilet debug",
-    "build": "pilet build",
-    "upgrade": "pilet upgrade"
-  },
-  "main": "dist/index.js",
-  "files": [
-    "dist"
-  ],
-  "importmap": {
-    "imports": {},
-    "inherit": [
-      "my-app"
-    ]
-  },
-  "source": "src/index.tsx"
-}
