--- my-app/pilets/product-pilet/src/Products.tsx
@@ -1,13 +1,0 @@
-import * as React from 'react';
-
-export default () => {
-  return (
-    <>
-      <h1>Products</h1>      
-      <p>
-        Lorem ipsum, dolor sit amet consectetur adipisicing elit. Fugiat reiciendis illum qui nulla adipisci, laudantium
-        vitae atque dicta blanditiis ab. Quae deleniti eum sapiente temporibus tenetur, maxime quibusdam iure nisi.
-      </p>
-    </>
-  );
-};
