--- my-app/pilets/product-pilet/tsconfig.json
@@ -1,25 +1,0 @@
-{
-  "compilerOptions": {
-    "declaration": true,
-    "noImplicitAny": false,
-    "removeComments": false,
-    "noLib": false,
-    "emitDecoratorMetadata": true,
-    "experimentalDecorators": true,
-    "target": "es6",
-    "sourceMap": true,
-    "outDir": "./dist",
-    "skipLibCheck": true,
-    "lib": ["dom", "es2018"],
-    "moduleResolution": "node",
-    "module": "esnext",
-    "jsx": "react",
-    "importHelpers": true
-  },
-  "include": [
-    "src"
-  ],
-  "exclude": [
-    "node_modules"
-  ]
-}
