+++ my-pilet/package.json
@@ -1,6 +1,6 @@
 {
   "name": "my-pilet",
-  "version": "1.1.3",
+  "version": "1.0.0",
   "description": "",
   "keywords": [
     "pilet"
@@ -9,9 +9,9 @@
     "@types/node": "latest",
     "@types/react": "^18",
     "@types/react-dom": "^18",
-    "my-app": "file:..\\dist\\emulator\\my-app-1.0.0.tgz",
-    "piral-cli": "1.2.0",
-    "piral-cli-webpack5": "1.2.0",
+    "app-shell": "file:../app-shell/dist/emulator/app-shell-1.0.0.tgz",
+    "piral-cli": "1.5.2",
+    "piral-cli-webpack5": "1.5.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-router": "5.3.4",
@@ -22,20 +22,18 @@
   "scripts": {
     "start": "pilet debug",
     "build": "pilet build",
-    "upgrade": "pilet upgrade"
+    "upgrade": "pilet upgrade",
+    "postinstall": "pilet declaration"
   },
   "main": "dist/index.js",
   "files": [
     "dist"
   ],
   "importmap": {
-    "imports": {},
     "inherit": [
-      "my-app"
-    ]
+      "app-shell"
+    ],
+    "imports": {}
   },
-  "source": "src/index.tsx",
-  "dependencies": {
-    "@liteobject/my-demo-npm": "^1.0.1"
-  }
+  "source": "src/index.tsx"
 }
