+++ my-pilet/pilet.json
@@ -1,7 +1,9 @@
 {
-  "schemaVersion": "v2",
   "$schema": "https://docs.piral.io/schemas/pilet-v0.json",
+  "schemaVersion": "v2",
   "piralInstances": {
-    "my-app": {}
+    "app-shell": {
+      "selected": true
+    }
   }
 }
